backend/monolito1/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
confluent-kafka==2.10.1
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-ckeditor-5==0.2.18
django-ckeditor==6.7.3 
django-cors-headers==4.7.0
django-environ==0.12.0
django-js-asset==3.1.2
django-pwa==2.0.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework-api-response==0.1.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
oauthlib==3.3.1
pillow==11.2.1
pika==1.3.2
pycparser==2.22
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.25.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0